Fix undefined index
Currently, when retrieving the KOM settings, we only check if KP exists when retrieving the KP settings even if it is not enabled (but is activated) although it is KCO that is enabled and is in use.  This leads to undefined index access since get_option returns `false`.

Check if we get back any settings from KP, otherwise, default to KCO settings.