Handle cart total change
Previously, when customers changed the shipping option, the `#kosm_cart_total` was unset, resulting in a NaN value when fetching the price.

With this pull request (PR), this issue is addressed by retrieving the cart total directly from the server using AJAX when it is modified. Also added safeguards to prevent passing NaN values to KOSM.